Se creó componente Contador.svelte
