+++ src/main.js
@@ -2,10 +2,6 @@ import App from './App.svelte';
 
 const app = new App({
 	target: document.body,
-	props: {
-		name: 'Desarrollo',
-		inicialConteo:5
-	}
 });
 
 export default app;
